requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
backports.tarfile==1.2.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
crashtest==0.4.1
distlib==0.3.9
distro==1.9.0
dulwich==0.22.8
fastapi==0.115.13
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
frozenlist==1.7.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
litellm==1.72.6.post2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
msgpack==1.1.1
multidict==6.5.0
openai==1.90.0
packaging==25.0
pbs-installer==2025.6.12
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry==2.1.3
poetry-core==2.1.3
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
Pygments==2.19.2
pyproject_hooks==1.2.0
python-dotenv==1.1.0
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
# rich==14.0.0
rpds-py==0.25.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
structlog==25.4.0
tavily-python==0.7.7
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.3
tqdm==4.67.1
trove-classifiers==2025.5.9.12
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
virtualenv==20.31.2
xattr==1.1.4
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
